requirements.txt
selenium==4.20.0
pillow
numpy
opencv-python
pytesseract
